Gestion des stocks + commentaire
La gestion des stocks est terminée

+++ js/app.js
@@ -1,125 +1,5 @@
-// Fonction qui parcours tout le tableau
-COURS.forEach(function(cours) {
-
-    // Fonction qui affiche les articles complétement en JS
-    creationCarteCours(cours.img, cours.titre, cours.note, cours.prix_initial, cours.prix, cours.stock, cours.id);
-
-});
-
-// Fonction qui crée toutes les cartes des cours disponible en JS grâce au tableau JSON cours
-function creationCarteCours(img, titre, note, prix_initial, prix, stock, id) {
-
-    // J'ai récréer tout l'ancien html avec les mêmes classes identiques
-    let parent = document.querySelector("body .container__cours");
-
-    let cours__article = document.createElement("div");
-
-    cours__article.classList.add("cours__article");
-
-    parent.appendChild(cours__article);
-
-    let cours_img = document.createElement("figure");
-
-    cours_img.classList.add("cours_img");
-
-    cours__article.appendChild(cours_img);
-
-    let cours_img_img = document.createElement("img");
-
-    cours_img_img.src = "img/cours/" + img;
-
-    cours_img.appendChild(cours_img_img);
-
-    let info__article = document.createElement("div");
-
-    info__article.classList.add("info__article");
-
-    cours__article.appendChild(info__article);
-
-    let h4 = document.createElement("h4");
-
-    h4.innerText = titre;
-
-    info__article.appendChild(h4);
-
-    let note_figure = document.createElement("figure");
-
-    note_figure.classList.add("note");
-
-    note_figure.classList.add("m_" + note);
-
-    info__article.appendChild(note_figure);
-
-    let note_img = document.createElement("img");
-
-    note_img.src = "img/rates.png";
-
-    note_figure.appendChild(note_img);
-
-    let paragrapheUn = document.createElement("p");
-
-    info__article.appendChild(paragrapheUn);
-
-    let spanParagrapheUn = document.createElement("span");
-
-    spanParagrapheUn.classList.add("prix");
-
-    spanParagrapheUn.innerText = prix_initial + " €";
-
-    paragrapheUn.appendChild(spanParagrapheUn);
-
-    let spanParagrapheUnDeux = document.createElement("span");
-
-    spanParagrapheUnDeux.classList.add("soldes");
-
-    spanParagrapheUnDeux.innerText = prix + " €";
-
-    paragrapheUn.appendChild(spanParagrapheUnDeux);
-
-    let paragrapheDeux = document.createElement("p");
-
-    paragrapheDeux.innerText = "Disponible: ";
-
-    info__article.append(paragrapheDeux);
-
-    let spanParagrapheDeux = document.createElement("span");
-
-    spanParagrapheDeux.classList.add("stock");
-
-    spanParagrapheDeux.innerText = stock;
-
-    paragrapheDeux.appendChild(spanParagrapheDeux);
-
-    let ajouter_au_panierA = document.createElement("a");
-
-    ajouter_au_panierA.href = "#";
-
-    ajouter_au_panierA.classList.add("ajouter-au-panier");
-
-    ajouter_au_panierA.setAttribute("data-id", id);
-
-    info__article.appendChild(ajouter_au_panierA);
-
-    let ajouter_au_panierI = document.createElement("i");
-
-    ajouter_au_panierI.classList.add("fa");
-
-    ajouter_au_panierI.classList.add("fa-cart-plus");
-
-    ajouter_au_panierA.appendChild(ajouter_au_panierI);
-
-    let spanAjouter_au_panier = document.createElement("span");
-
-    spanAjouter_au_panier.innerText = "Ajouter au panier";
-
-    ajouter_au_panierA.appendChild(spanAjouter_au_panier);
-
-}
-
 // On stock et on récupère le local storage
-let Objet = getLocalStorage();
-
-console.log(Objet)
+let Objet = recupererLeStorage();
 
 // Si la taille du tableu est différent de null alors on boucle sur les article, et on les affiche
 if (Objet.length != null) {
@@ -143,45 +23,55 @@ panier.forEach(function(cours) {
         // On récupère l'id grâce à l'attribut data-id
         let idCours = cours.getAttribute("data-id");
 
-        // On remonte à l'élément parent du bouton en question
-        let parentPanier = event.target;
+        // On récupérer le stock
+        let stockCours = recupererLeStorageCours();
 
-        // Vu que l'on a rajouter un span dans le a on est obligé rechercher le parent du parent
-        let parentPanierElement = parentPanier.parentElement.parentElement;
+        if (stockCours[idCours - 1].stock > 0) {
+            // On remonte à l'élément parent du bouton en question
+            let parentPanier = event.target;
 
-        // On récupére le nom du cours
-        let contenuPanier = parentPanierElement.querySelector("h4");
+            // Vu que l'on a rajouter un span dans le a on est obligé rechercher le parent du parent
+            let parentPanierElement = parentPanier.parentElement.parentElement;
 
-        // On récupére le prix en solde
-        let prixCours = parentPanierElement.querySelector(".soldes");
+            // On récupére le nom du cours
+            let contenuPanier = parentPanierElement.querySelector("h4");
 
-        // On récupérer le solde
-        let stockCours = parentPanierElement.querySelector(".stock")
+            // On récupére le prix en solde
+            let prixCours = parentPanierElement.querySelector(".soldes");
 
-        // Je défini le tableau grâce à la fonction qui récupère le local storage
-        let Objet = getLocalStorage();
+            // Je défini le tableau grâce à la fonction qui récupère le local storage
+            let Objet = recupererLeStorage();
 
-        // On push à chaque clique dans le local storage
-        Objet.push({
-            id: idCours,
-            nom: contenuPanier.textContent,
-            prix: prixCours.textContent,
-            quantite: 1
-        })
+            // On push à chaque clique dans le local storage
+            Objet.push({
+                id: idCours,
+                nom: contenuPanier.textContent,
+                prix: prixCours.textContent,
+                quantite: 1
+            });
 
-        ajoutCoursPanier(Objet[Objet.length - 1].nom, Objet[Objet.length - 1].prix, Objet[Objet.length - 1].quantite, Objet[Objet.length - 1].id);
+            ajoutCoursPanier(Objet[Objet.length - 1].nom, Objet[Objet.length - 1].prix, Objet[Objet.length - 1].quantite, Objet[Objet.length - 1].id);
 
-        // Ca ajoute le tableau objet au local storage
-        localStorage.setItem("panier", JSON.stringify(Objet));
+            // Ca ajoute le tableau objet au local storage
+            localStorage.setItem("panier", JSON.stringify(Objet));
 
-        //stockCours.innerText = stockCours.value - 1;
+            // Désincrémenté le code
+            stockCours[idCours - 1].stock = stockCours[idCours - 1].stock - 1;
 
+            // Actualise le html
+            parentPanierElement.querySelector(".stock").innerText = stockCours[idCours - 1].stock;
+
+            // Actualise le local storage
+            localStorage.setItem("cours", JSON.stringify(stockCours));
+
+        } else {
+            confirm('stock vide');
+        }
     })
-
 });
 
 // Récupére le local storage
-function getLocalStorage() {
+function recupererLeStorage() {
 
     // Stock les données récupérer du local storage panier
     let recupCoursPanier = localStorage.getItem("panier");
@@ -266,21 +156,47 @@ panierContainer.addEventListener("click", function(event) {
     // On vérifie que l'élement selectionné dans cette div comporte bien la class "supprimer-du-panier",
     if (event.target.className == "supprimer-du-panier") {
 
-        // On récup le local storage dans le tableau
-        let Objet = getLocalStorage();
-
         // On remonte au parent parent de l'élément
         let panierTable = event.target.parentElement.parentElement;
 
-        // On le supprimer grâce à l'id en cours
-        Objet.splice(Objet.id, 1);
+        // récupère le nom de l'article
+        let nom = panierTable.getElementsByTagName("td")[1].textContent;
 
-        // On vide l'html du panier
-        panierTable.innerHTML = "";
+        // Ouvre une boite de dialogue qui demande, si elle clique sur OK
+        if (confirm("Êtes vous sur de vouloir supprimer le cours " + nom + " ?")) {
 
-        // On actualise le local storage
-        localStorage.setItem("panier", JSON.stringify(Objet));
+            // On récup le local storage dans le tableau
+            let Objet = recupererLeStorage();
 
+            let idCours = event.target.getAttribute("data-id");
+
+            // On le supprimer grâce à l'id en cours grâce à la boucle
+            for (let i = 0; i < Objet.length; i++) {
+                if (Objet[i].id === idCours) {
+                    Objet.splice(i, 1);
+                    break;
+                }
+            }
+
+            // On vide l'html du panier
+            panierTable.innerHTML = '';
+
+            // On récupérer le stock
+            let stockCours = recupererLeStorageCours();
+
+            stockCours[idCours - 1].stock = stockCours[idCours - 1].stock++;
+
+            // document.querySelector(".stock").innerText = stockCours[idCours - 1].stock;
+            document.querySelector("body .container__cours").innerHTML = '';
+            affichage(stockCours);
+
+            // On actualise le stock
+            localStorage.setItem("cours", JSON.stringify(stockCours));
+
+            // On actualise le local storage
+            localStorage.setItem("panier", JSON.stringify(Objet));
+
+        }
     }
 
 })
@@ -288,13 +204,20 @@ panierContainer.addEventListener("click", function(event) {
 // Lors du clic sur le bouton vider le panier
 document.getElementById("vider-panier").addEventListener("click", function() {
 
-    // On selectionne le tbody du panier
-    let panierTable = document.querySelector("#table-panier tbody");
+    // Ouvre une boite de dialogue qui demande, si elle clique sur OK
+    if (confirm("Voulez vous vraiment vider votre panier ?")) {
 
-    // Vide le html
-    panierTable.innerHTML = "";
+        // On selectionne le tbody du panier
+        let panierTable = document.querySelector("#table-panier tbody");
 
-    // Le local storage est clear
-    localStorage.removeItem("panier");
+        // Vide le html
+        panierTable.innerHTML = "";
+
+        // Le local storage est clear
+        localStorage.clear;
+
+        alert("N'hésitez pas à remplir votre panier, si les cours vous plaisent ;) !")
+
+    }
 
 });
